Chore: Move `parser-babel-config` to a package
