+++ packages/parser-babel-config/CHANGELOG.md
@@ -1,0 +1,10 @@
+# 1.0.1 (March 7, 2018)
+
+## Bug fixes / Improvements
+
+* Optimize published package.
+
+
+# 1.0.0 (March 6, 2018)
+
+✨
