+++ packages/parser-babel-config/package.json
@@ -7,20 +7,21 @@
     "timeout": "1m"
   },
   "dependencies": {
-    "ajv": "^6.1.1"
+    "ajv": "^6.2.1"
   },
-  "description": "",
+  "description": "sonarwhal parser needed to analyze babel config",
   "devDependencies": {
     "ava": "^0.25.0",
     "cpx": "^1.5.0",
-    "eslint": "^4.18.1",
+    "eslint": "^4.18.2",
     "eslint-plugin-import": "^2.8.0",
     "eslint-plugin-markdown": "^1.0.0-beta.7",
-    "eslint-plugin-typescript": "^0.8.1",
+    "eslint-plugin-typescript": "^0.9.0",
     "markdownlint-cli": "^0.7.0",
     "npm-link-check": "^2.0.0",
     "npm-run-all": "^4.1.2",
     "nyc": "^11.4.1",
+    "proxyquire": "^2.0.0",
     "rimraf": "^2.6.2",
     "sonarwhal": "^0.29.0-beta.0",
     "typescript": "^2.6.2",
@@ -36,35 +37,33 @@
   "homepage": "https://sonarwhal.com/",
   "keywords": [
     "sonarwhal",
-    "sonarwhal-rule",
-    "babel-config-is-valid",
-    "babel-config-is-valid-rule",
-    "sonarwhal-recommended"
+    "sonarwhal-parser",
+    "babel-config"
   ],
   "license": "Apache-2.0",
   "main": "dist/src/index.js",
-  "name": "@sonarwhal/rule-babel-config-is-valid",
+  "name": "@sonarwhal/parser-babel-config",
   "nyc": {
     "extends": "../../../../.nycrc"
   },
   "peerDependencies": {
     "sonarwhal": "^0.29.0-beta.0"
   },
-  "private": true,
   "repository": "sonarwhal/sonarwhal",
   "scripts": {
-    "build": "npm run clean && npm-run-all build:*",
+    "build": "yarn clean && npm-run-all build:*",
+    "build-release": "yarn clean && yarn build:assets && tsc --inlineSourceMap false --removeComments true",
     "build:assets": "cpx \"./{src,tests}/**/{!(*.ts),.!(ts)}\" dist",
     "build:ts": "tsc",
     "clean": "rimraf dist",
     "lint": "npm-run-all lint:*",
-    "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path ../../.eslintignore --report-unused-disable-directives --ignore-pattern dist/**/*.js --ignore-pattern dist/**/*.d.ts",
-    "lint:md": "markdownlint *.md",
-    "test": "npm run lint && npm run build && nyc ava",
-    "watch": "npm run build && npm-run-all --parallel -c watch:*",
-    "watch:assets": "npm run build:assets -- -w --no-initial",
+    "lint:js": "eslint . --cache --ext js --ext md --ext ts --ignore-path ../../.eslintignore --report-unused-disable-directives",
+    "lint:md": "markdownlint --ignore CHANGELOG.md *.md",
+    "test": "yarn lint && yarn build && nyc ava",
+    "watch": "yarn build && npm-run-all --parallel -c watch:*",
+    "watch:assets": "yarn build:assets -- -w --no-initial",
     "watch:test": "ava --watch",
-    "watch:ts": "npm run build:ts -- --watch"
+    "watch:ts": "yarn build:ts -- --watch"
   },
-  "version": "1.0.0"
+  "version": "1.0.2-beta.0"
 }
