+++ packages/parser-babel-config/src/index.ts
@@ -1,0 +1,1 @@
+module.exports = { 'babel-config': require('./babel-config') };
