+++ packages/parser-babel-config/src/schema/babelConfigSchema.json
