+++ packages/parser-babel-config/tests/babel-config.ts
@@ -2,7 +2,7 @@ import * as sinon from 'sinon';
 import test from 'ava';
 import { EventEmitter2 } from 'eventemitter2';
 
-import BabelConfigParser, * as babelConfigParser from '../../../src/lib/parsers/babel-config/babel-config';
+import BabelConfigParser from '../src/babel-config';
 
 test.beforeEach((t) => {
     t.context.sonarwhal = new EventEmitter2({
@@ -14,8 +14,8 @@ test.beforeEach((t) => {
 
 test.serial('If any file is parsed, it should emit a `notfound::babel-config` error', async (t) => {
     const sandbox = sinon.sandbox.create();
-    const parser = new BabelConfigParser(t.context.sonarwhal);
 
+    new BabelConfigParser(t.context.sonarwhal); // eslint-disable-line no-new
     sandbox.spy(t.context.sonarwhal, 'emitAsync');
 
     await t.context.sonarwhal.emitAsync('scan::end', {});
@@ -26,8 +26,8 @@ test.serial('If any file is parsed, it should emit a `notfound::babel-config` er
 
 test.serial(`If the resource doesn't match the regex, nothing should happen`, async (t) => {
     const sandbox = sinon.sandbox.create();
-    const parser = new BabelConfigParser(t.context.sonarwhal);
 
+    new BabelConfigParser(t.context.sonarwhal); // eslint-disable-line no-new
     sandbox.spy(t.context.sonarwhal, 'emitAsync');
 
     await t.context.sonarwhal.emitAsync('fetch::end::json', { resource: '.babelrcconfig' });
@@ -37,8 +37,8 @@ test.serial(`If the resource doesn't match the regex, nothing should happen`, as
 
 test.serial('If the file contains an invalid json, it should fail', async (t) => {
     const sandbox = sinon.sandbox.create();
-    const parser = new BabelConfigParser(t.context.sonarwhal);
 
+    new BabelConfigParser(t.context.sonarwhal); // eslint-disable-line no-new
     sandbox.spy(t.context.sonarwhal, 'emitAsync');
 
     await t.context.sonarwhal.emitAsync('fetch::end::json', {
@@ -52,7 +52,8 @@ test.serial('If the file contains an invalid json, it should fail', async (t) =>
 
 test.serial('If the file contains an invalid schema, it should fail', async (t) => {
     const sandbox = sinon.sandbox.create();
-    const parser = new BabelConfigParser(t.context.sonarwhal);
+
+    new BabelConfigParser(t.context.sonarwhal); // eslint-disable-line no-new
     const invalidSchemaContent = `{
         "plugins": ["transform-react-jsx"],
         "moduleId": 1,
@@ -76,7 +77,8 @@ test.serial('If the file contains an invalid schema, it should fail', async (t) 
 
 test.serial('If the content type is unknown, it should still validate if the file name is a match', async (t) => {
     const sandbox = sinon.sandbox.create();
-    const parser = new BabelConfigParser(t.context.sonarwhal);
+
+    new BabelConfigParser(t.context.sonarwhal); // eslint-disable-line no-new
     const invalidSchemaContent = `{
         "plugins": ["transform-react-jsx"],
         "moduleId": 1,
@@ -97,5 +99,3 @@ test.serial('If the content type is unknown, it should still validate if the fil
     t.is(t.context.sonarwhal.emitAsync.args[1][0], 'invalid-schema::babel-config');
     t.is(t.context.sonarwhal.emitAsync.args[2][0], 'parse::babel-config');
 });
-
-
